+++ source/Examples/script.ps1
@@ -1,0 +1,3 @@
+import-module PSMSALNet
+
+Get-EntraToken -SystemManagedIdentity -Resource GraphAPI